fix(codegen): LIDL int/uint are 64-bit in the Qt spelling too
lidlTypeToQt mapped BOTH `int` and `uint` to plain `int`. Everywhere else in the
stack a LIDL int/uint is 64-bit — int64_t/uint64_t in C++ impls, i64/u64 in the
Rust SDK — so the Qt spelling broke the one-type-per-LIDL-type rule and lost
data: a Qt consumer reading a `uint` return got a SIGNED 32-bit value, so
anything above 2^31 came back wrong and anything above 2^63 was never
expressible.

int -> qlonglong, uint -> qulonglong, and returnConversion() gains the matching
accessors (toLongLong / toULongLong instead of toInt).

qlonglong/qulonglong rather than qint64/quint64 so the generated introspection
JSON uses the same names Qt's own metaobject normalisation produces — otherwise
a cdylib module's generated `signature` and a legacy module's
QMetaObject-derived one would disagree for the same LIDL type. Nothing looks
these strings up: the only QMetaType::fromName call in the stack is for
"LogosResult".

This changes two generated surfaces: the Qt consumer wrapper signatures and the
introspection JSON. Passing an int argument still converts implicitly, so
callers keep compiling; code that assigns a wrapper's return into an `int`
narrows and may warn, which is the bug being surfaced rather than a regression.

Tests: 168/168, with the type-mapping and client-emitter expectations updated to
the 64-bit spelling.

+++ cpp-generator/experimental/lidl_emit_common.cpp
@@ -20,8 +20,14 @@ QString lidlTypeToQt(const TypeExpr& te)
         if (te.name == "void")    return "void";
         if (te.name == "tstr")    return "QString";
         if (te.name == "bstr")    return "QByteArray";
-        if (te.name == "int")     return "int";
-        if (te.name == "uint")    return "int";
+        // 64-bit, and unsigned stays unsigned. LIDL int/uint are int64_t/uint64_t
+        // everywhere else (C++ impls, Rust's i64/u64), so spelling them `int`
+        // here broke the 1-1 mapping and truncated: a Qt consumer reading a
+        // `uint` return got a SIGNED 32-bit value. qlonglong/qulonglong rather
+        // than qint64/quint64 so the generated introspection matches the names
+        // Qt's own metaobject normalisation produces.
+        if (te.name == "int")     return "qlonglong";
+        if (te.name == "uint")    return "qulonglong";
         if (te.name == "float64") return "double";
         if (te.name == "bool")    return "bool";
         if (te.name == "result")  return "LogosResult";
